+++ EventListener/FlashListener.php
@@ -12,7 +12,6 @@
 namespace FOS\UserBundle\EventListener;
 
 use FOS\UserBundle\FOSUserEvents;
-use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
@@ -23,11 +22,6 @@ class FlashListener implements EventSubscriberInterface
      * @var string[]
      */
     private static $successMessages = [
-        FOSUserEvents::CHANGE_PASSWORD_COMPLETED => 'change_password.flash.success',
-        FOSUserEvents::GROUP_CREATE_COMPLETED => 'group.flash.created',
-        FOSUserEvents::GROUP_DELETE_COMPLETED => 'group.flash.deleted',
-        FOSUserEvents::GROUP_EDIT_COMPLETED => 'group.flash.updated',
-        FOSUserEvents::PROFILE_EDIT_COMPLETED => 'profile.flash.updated',
         FOSUserEvents::REGISTRATION_COMPLETED => 'registration.flash.user_created',
         FOSUserEvents::RESETTING_RESET_COMPLETED => 'resetting.flash.success',
     ];
@@ -57,11 +51,6 @@ class FlashListener implements EventSubscriberInterface
     public static function getSubscribedEvents()
     {
         return [
-            FOSUserEvents::CHANGE_PASSWORD_COMPLETED => 'addSuccessFlash',
-            FOSUserEvents::GROUP_CREATE_COMPLETED => 'addSuccessFlash',
-            FOSUserEvents::GROUP_DELETE_COMPLETED => 'addSuccessFlash',
-            FOSUserEvents::GROUP_EDIT_COMPLETED => 'addSuccessFlash',
-            FOSUserEvents::PROFILE_EDIT_COMPLETED => 'addSuccessFlash',
             FOSUserEvents::REGISTRATION_COMPLETED => 'addSuccessFlash',
             FOSUserEvents::RESETTING_RESET_COMPLETED => 'addSuccessFlash',
         ];
@@ -70,7 +59,7 @@ class FlashListener implements EventSubscriberInterface
     /**
      * @param string $eventName
      */
-    public function addSuccessFlash(Event $event, $eventName)
+    public function addSuccessFlash($event, $eventName)
     {
         if (!isset(self::$successMessages[$eventName])) {
             throw new \InvalidArgumentException('This event does not correspond to a known flash message');