+++ Tests/Routing/RoutingTest.php
@@ -30,13 +30,6 @@ class RoutingTest extends TestCase
         $loader = new XmlFileLoader($locator);
 
         $collection = new RouteCollection();
-        $collection->addCollection($loader->load(__DIR__.'/../../Resources/config/routing/change_password.xml'));
-        $subCollection = $loader->load(__DIR__.'/../../Resources/config/routing/group.xml');
-        $subCollection->addPrefix('/group');
-        $collection->addCollection($subCollection);
-        $subCollection = $loader->load(__DIR__.'/../../Resources/config/routing/profile.xml');
-        $subCollection->addPrefix('/profile');
-        $collection->addCollection($subCollection);
         $subCollection = $loader->load(__DIR__.'/../../Resources/config/routing/registration.xml');
         $subCollection->addPrefix('/register');
         $collection->addCollection($subCollection);
@@ -57,16 +50,7 @@ class RoutingTest extends TestCase
     public function loadRoutingProvider()
     {
         return [
-            ['fos_user_change_password', '/change-password', ['GET', 'POST']],
-
-            ['fos_user_group_list', '/group/list', ['GET']],
-            ['fos_user_group_new', '/group/new', ['GET', 'POST']],
-            ['fos_user_group_show', '/group/{groupName}', ['GET']],
-            ['fos_user_group_edit', '/group/{groupName}/edit', ['GET', 'POST']],
-            ['fos_user_group_delete', '/group/{groupName}/delete', ['GET']],
-
             ['fos_user_profile_show', '/profile/', ['GET']],
-            ['fos_user_profile_edit', '/profile/edit', ['GET', 'POST']],
 
             ['fos_user_registration_register', '/register/', ['GET', 'POST']],
             ['fos_user_registration_check_email', '/register/check-email', ['GET']],